Move enum classes into separate file
